requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
librosa>=0.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.22.0
Pillow>=10.0.0
pyyaml>=6.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

